+++ src/pages/eunjungSung/Main/Components/Feed.scss
@@ -1,0 +1,63 @@
+@import '../../../../styles/mixin.scss';
+
+.feed {
+  .profileWrapper {
+    display: flex;
+    justify-content: space-between;
+    padding: 10px;
+    text-align: center;
+
+    .profileLeft {
+      @include flex(center, center);
+      gap: 10px;
+    }
+    .menuImg {
+      width: 16px;
+      height: 16px;
+    }
+  }
+
+  .articleImg {
+    width: 454px;
+    height: 454px;
+  }
+
+  .functionWrapper {
+    @include flex(space-between, center);
+    padding: 10px;
+
+    .heartBtn {
+      @include flex(center, center);
+      border: none;
+      margin: 0px;
+      padding: 0px;
+      background-color: white;
+      cursor: pointer;
+
+      .heartBtnImg {
+        width: 24px;
+        height: 24px;
+      }
+    }
+
+    .functionRight {
+      display: flex;
+      gap: 10px;
+    }
+  }
+
+  .likeWrapper {
+    @include flex(left, center);
+    text-align: center;
+  }
+
+  .likeImg {
+    margin: 0px 10px;
+  }
+
+  .descWrapper {
+    display: flex;
+    gap: 10px;
+    margin: 10px;
+  }
+}
